requirements.txt
Django==4.1
djangorestframework==3.13.1
python-dotenv==0.20.0
pytest-django==4.5.2
pytest-cov==3.0.0
PyJWT==2.4.0
